fix: exception handling in ```GET/PATCH/DELETE``` req
(glasses.controller.ts): add swagger documentation for api response w/ exception
(glasses.service.ts): use DRY principles

closed #3

+++ src/glasses/glasses.service.ts
@@ -1,11 +1,15 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 
 import { CreateGlassDto } from './dto/create-glass.dto';
 import { UpdateGlassDto } from './dto/update-glass.dto';
+import { FindOneGlassDto } from './dto/find-glass.dto';
 
 import { Glass } from './entities/glass.entity';
-import { FindOneGlassDto } from './dto/find-glass.dto';
 
 @Injectable()
 export class GlassesService {
@@ -13,18 +17,12 @@ export class GlassesService {
 
   async create(createGlassDto: CreateGlassDto): Promise<Glass> {
     const { name, vendor_code } = createGlassDto;
-    const existedGlassName = await this.findOne({ name });
+    const check = await this.checkNameAndCode(name, vendor_code);
 
-    if (existedGlassName)
-      throw new ConflictException('Объект с таким именем уже существует');
-
-    const existedVendorCode = await this.findOne({ vendor_code });
-
-    if (existedVendorCode)
-      throw new ConflictException('Объект с таким vendor_code уже существует');
-
-    const newGlass = await this.glassModule.create({ ...createGlassDto });
-    return newGlass;
+    if (check) {
+      const newGlass = await this.glassModule.create({ ...createGlassDto });
+      return newGlass;
+    }
   }
 
   async findAll(): Promise<Glass[]> {
@@ -35,15 +33,46 @@ export class GlassesService {
     return await this.glassModule.findOne({ where: { ...props } });
   }
 
-  async update(id: number, updateGlassDto: UpdateGlassDto): Promise<Glass> {
-    await this.glassModule.update({ ...updateGlassDto }, { where: { id } });
-    const updatedGalss = await this.glassModule.findByPk(id);
-    return updatedGalss;
+  async update(id: number, updateGlassDto: UpdateGlassDto) {
+    const { name, vendor_code } = updateGlassDto;
+    const existedGlass = await this.findOne({ id });
+
+    if (!existedGlass) throw new NotFoundException('объект не найден');
+
+    const check = await this.checkNameAndCode(name, vendor_code, id);
+
+    if (check) {
+      await this.glassModule.update({ ...updateGlassDto }, { where: { id } });
+      const updatedGalss = await this.glassModule.findByPk(id);
+      return updatedGalss;
+    }
   }
 
   async remove(id: number): Promise<Glass> {
     const removed = await this.glassModule.findByPk(id);
+
+    if (!removed) throw new NotFoundException('объект не найден');
+
     await this.glassModule.destroy({ where: { id } });
     return removed;
+  }
+
+  private async checkNameAndCode(
+    name: string,
+    vendor_code: string,
+    id?: number,
+  ): Promise<boolean> {
+    const existedGlassName = await this.findOne({ name: name || null });
+    if (existedGlassName && existedGlassName.id !== id)
+      throw new ConflictException('Объект с таким именем уже существует');
+
+    const existedVendorCode = await this.findOne({
+      vendor_code: vendor_code || null,
+    });
+
+    if (existedVendorCode && existedVendorCode.id !== id)
+      throw new ConflictException('Объект с таким vendor_code уже существует');
+
+    return true;
   }
 }
